refactor: Let K be of type u32
This allows to get rid of some of the numerical warnings. It does not
have to be casted from 64bit no 32bit anymore.

+++ src/main.rs
@@ -15,7 +15,7 @@ mod utils;
 
 fn main() {
     const N: usize = 7;
-    const K: usize = 3;
+    const K: u32 = 3;
     let mut mcts_referee = NaiveReferee::<N, K> {};
     // let mut mcts_base_player0 = OneLookaheadPlayer::new(1, Box::new(NaiveReferee::<N, K> {}), 0);
     // let mut mcts_base_player1 = OneLookaheadPlayer::new(0, Box::new(NaiveReferee::<N, K> {}), 1);
